Rename PackageInfoSection -> AdvancedConfigurationSection
