--- wi/wi-webviews/src/views/creationView/biForm/components/PackageInfoSection.tsx
@@ -1,236 +1,0 @@
-/**
- * Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com) All Rights Reserved.
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import { Codicon, Dropdown, TextField } from "@wso2/ui-toolkit";
-import {
-    Description,
-    FieldGroup,
-    Note,
-    SubSectionDivider,
-    SubSectionLabel,
-    SignInHint,
-    SignInHintButton
-} from "../styles";
-import { CollapsibleSection } from "./CollapsibleSection";
-import { sanitizePackageName, sanitizeProjectHandle } from "../utils";
-import { useSignIn } from "../../../../hooks/useSignIn";
-
-export interface PackageInfoData {
-    packageName: string;
-    orgName: string;
-    version: string;
-    projectHandle?: string;
-}
-
-export interface Organization {
-    id?: number | string;
-    handle: string;
-    name: string;
-}
-
-export interface PackageInfoSectionProps {
-    /** Whether the section is expanded */
-    isExpanded: boolean;
-    /** Callback when the section is toggled */
-    onToggle: () => void;
-    /** The package info data */
-    data: PackageInfoData;
-    /** Callback when the package info changes */
-    onChange: (data: Partial<PackageInfoData>) => void;
-    /** Whether the package is a library */
-    isLibrary?: boolean;
-    /** Error message for org name validation */
-    orgNameError?: string | null;
-    /** Error message for package name validation */
-    packageNameError?: string | null;
-    /** Error message for project handle validation */
-    projectHandleError?: string | null;
-    /** Organizations list — when provided, renders a dropdown instead of a free-text field */
-    organizations?: Organization[];
-    /** Whether the section contains validation errors */
-    hasError?: boolean;
-}
-
-// ── Component ──────────────────────────────────────────────────────────────────
-
-export function PackageInfoSection({
-    isExpanded,
-    onToggle,
-    data,
-    onChange,
-    isLibrary,
-    orgNameError,
-    packageNameError,
-    projectHandleError,
-    organizations,
-    hasError
-}: PackageInfoSectionProps) {
-    const { isSigningIn, handleSignIn, handleCancelSignIn } = useSignIn();
-
-    const hasOrgs = organizations && organizations.length > 0;
-    const createWithinProject = data.projectHandle !== undefined;
-
-    return (
-        <CollapsibleSection
-            isExpanded={isExpanded}
-            onToggle={onToggle}
-            icon="gear"
-            title="Advanced Configurations"
-            hasError={hasError}
-        >
-            {createWithinProject && (
-                <>
-                    <SubSectionLabel>Project</SubSectionLabel>
-                    <FieldGroup>
-                    {hasOrgs ? (
-                            <>
-                                <Dropdown
-                                    id="org-name-dropdown"
-                                    label="Organization Name"
-                                    items={organizations.map((org) => ({ value: org.handle, content: org.name }))}
-                                    value={data.orgName}
-                                    onValueChange={(value: string) => onChange({ orgName: value })}
-                                />
-                                <Description>The organization that owns this project.</Description>
-                            </>
-                        ) : (
-                            <>
-                                <TextField
-                                    onTextChange={(value) => onChange({ orgName: value })}
-                                    value={data.orgName}
-                                    label="Organization Name"
-                                    errorMsg={orgNameError || undefined}
-                                />
-                                <SignInHint>
-                                    <Codicon
-                                        name="account"
-                                        iconSx={{ color: "var(--vscode-descriptionForeground)" }}
-                                        sx={{ display: "flex" }}
-                                    />
-                                    <span>Sign in to pick from your organizations —</span>
-                                    <SignInHintButton type="button" onClick={isSigningIn ? handleCancelSignIn : handleSignIn}>
-                                        {isSigningIn ? (
-                                            <>
-                                                <Codicon
-                                                    name="loading"
-                                                    iconSx={{ fontSize: "11px", animation: "codicon-spin 1.5s steps(30) infinite" }}
-                                                />
-                                                Signing in...
-                                                <Codicon
-                                                    name="close"
-                                                    iconSx={{ fontSize: "10px" }}
-                                                />
-                                            </>
-                                        ) : (
-                                            "Sign In"
-                                        )}
-                                    </SignInHintButton>
-                                </SignInHint>
-                            </>
-                        )
-                    }
-                </FieldGroup>
-                    <FieldGroup>
-                        <TextField
-                            onTextChange={(value) => onChange({ projectHandle: sanitizeProjectHandle(value, { trimTrailing: false }) })}
-                            value={data.projectHandle}
-                            label="Project ID"
-                            errorMsg={projectHandleError || undefined}
-                        />
-                        <Description>Unique identifier for your project in various contexts. Cannot be changed after creation.</Description>
-                    </FieldGroup>
-                    <SubSectionDivider />
-                </>
-            )}
-            <SubSectionLabel>Ballerina Package</SubSectionLabel>
-            <Note style={{ marginBottom: "16px" }}>
-                {createWithinProject
-                    ? `This ${isLibrary ? "library" : "integration"} is generated as a Ballerina package. Specify the package name and version to be assigned.`
-                    : `This ${isLibrary ? "library" : "integration"} is generated as a Ballerina package. Specify the organization, package name and version to be assigned.`}
-            </Note>
-            <FieldGroup>
-                <TextField
-                    onTextChange={(value) => onChange({ packageName: sanitizePackageName(value) })}
-                    value={data.packageName}
-                    label="Package Name"
-                    errorMsg={packageNameError || undefined}
-                />
-                <Description>Specify the package name.</Description>
-            </FieldGroup>
-            {!createWithinProject && (
-                <FieldGroup>
-                    {hasOrgs ? (
-                        <>
-                            <Dropdown
-                                id="org-name-dropdown"
-                                label="Organization Name"
-                                items={organizations.map((org) => ({ value: org.handle, content: org.name }))}
-                                value={data.orgName}
-                                onValueChange={(value: string) => onChange({ orgName: value })}
-                            />
-                            <Description>The organization that owns this package.</Description>
-                        </>
-                    ) : (
-                        <>
-                            <TextField
-                                onTextChange={(value) => onChange({ orgName: value })}
-                                value={data.orgName}
-                                label="Organization Name"
-                                errorMsg={orgNameError || undefined}
-                            />
-                            <SignInHint>
-                                <Codicon
-                                    name="account"
-                                    iconSx={{ color: "var(--vscode-descriptionForeground)" }}
-                                    sx={{ display: "flex" }}
-                                />
-                                <span>Sign in to pick from your organizations —</span>
-                                <SignInHintButton type="button" onClick={isSigningIn ? handleCancelSignIn : handleSignIn}>
-                                    {isSigningIn ? (
-                                        <>
-                                            <Codicon
-                                                name="loading"
-                                                iconSx={{ fontSize: "11px", animation: "codicon-spin 1.5s steps(30) infinite" }}
-                                            />
-                                            Signing in...
-                                            <Codicon
-                                                name="close"
-                                                iconSx={{ fontSize: "10px" }}
-                                            />
-                                        </>
-                                    ) : (
-                                        "Sign In"
-                                    )}
-                                </SignInHintButton>
-                            </SignInHint>
-                        </>
-                    )}
-                </FieldGroup>
-            )}
-            <FieldGroup>
-                <TextField
-                    onTextChange={(value) => onChange({ version: value })}
-                    value={data.version}
-                    label="Package Version"
-                    placeholder="0.1.0"
-                />
-                <Description>Version of the package.</Description>
-            </FieldGroup>
-        </CollapsibleSection>
-    );
-}
