+++ wi/wi-webviews/src/views/creationView/biForm/components/index.ts
@@ -19,5 +19,5 @@
 export { CollapsibleSection } from "./CollapsibleSection";
 export type { CollapsibleSectionProps } from "./CollapsibleSection";
 
-export { PackageInfoSection } from "./PackageInfoSection";
-export type { PackageInfoSectionProps, PackageInfoData } from "./PackageInfoSection";
+export { AdvancedConfigurationSection } from "./AdvancedConfigurationSection";
+export type { AdvancedConfigurationSectionProps, ConfigurationData } from "./AdvancedConfigurationSection";
